requirements.txt
fastmcp
pyyaml
questionary
